pyproject.toml
[project]
name = "todo-saas-back"
version = "0.1.0"
description = ""
authors = [
    {name = "Jose G",email = "josegarciama@hotmail.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "google-cloud-firestore (>=2.21.0,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
httpx = "^0.28.1"
ruff = "^0.12.12"
mypy = "^1.17.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
